Preliminary task: Moved enum ApiAction.
