+++ oapenwb-javalin/src/main/java/dk/ule/oapenwb/data/importer/csv/components/MappingMaker.java
@@ -6,7 +6,7 @@ import dk.ule.oapenwb.AdminControllers;
 import dk.ule.oapenwb.base.error.CodeException;
 import dk.ule.oapenwb.data.importer.csv.CsvImporterConfig;
 import dk.ule.oapenwb.data.importer.csv.CsvImporterContext;
-import dk.ule.oapenwb.entity.basis.ApiAction;
+import dk.ule.oapenwb.entity.ApiAction;
 import dk.ule.oapenwb.entity.content.basedata.LangPair;
 import dk.ule.oapenwb.entity.content.lexemes.Mapping;
 import lombok.Getter;
