+++ oapenwb-javalin/src/main/java/dk/ule/oapenwb/entity/ApiAction.java
@@ -1,6 +1,6 @@
 // SPDX-FileCopyrightText: © 2022 Michael Köther <mkoether38@gmail.com>
 // SPDX-License-Identifier: AGPL-3.0-only
-package dk.ule.oapenwb.entity.basis;
+package dk.ule.oapenwb.entity;
 
 /**
  * Specifies an action done to an entity (for some entities) via the REST interface when the REST interface
