+++ oapenwb-javalin/src/main/java/dk/ule/oapenwb/entity/content/lexemes/lexeme/Lexeme.java
@@ -4,7 +4,7 @@ package dk.ule.oapenwb.entity.content.lexemes.lexeme;
 
 import com.fasterxml.jackson.annotation.JsonView;
 import dk.ule.oapenwb.base.Views;
-import dk.ule.oapenwb.entity.basis.ApiAction;
+import dk.ule.oapenwb.entity.ApiAction;
 import dk.ule.oapenwb.logic.admin.lexeme.IRPCEntity;
 import io.hypersistence.utils.hibernate.type.json.JsonBinaryType;
 import jakarta.persistence.*;
