+++ oapenwb-javalin/src/main/java/dk/ule/oapenwb/entity/content/lexemes/lexeme/Variant.java
@@ -4,7 +4,7 @@ package dk.ule.oapenwb.entity.content.lexemes.lexeme;
 
 import com.fasterxml.jackson.annotation.JsonView;
 import dk.ule.oapenwb.base.Views;
-import dk.ule.oapenwb.entity.basis.ApiAction;
+import dk.ule.oapenwb.entity.ApiAction;
 import dk.ule.oapenwb.entity.content.lexemes.LexemeForm;
 import dk.ule.oapenwb.entity.content.lexemes.MetaInfo;
 import dk.ule.oapenwb.logic.admin.lexeme.IRPCEntity;
