+++ oapenwb-javalin/src/test/java/dk/ule/oapenwb/data/importer/VariantUtil.java
@@ -3,7 +3,7 @@
 package dk.ule.oapenwb.data.importer;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import dk.ule.oapenwb.entity.basis.ApiAction;
+import dk.ule.oapenwb.entity.ApiAction;
 import dk.ule.oapenwb.entity.content.lexemes.LexemeForm;
 import dk.ule.oapenwb.entity.content.lexemes.lexeme.Variant;
 import io.javalin.json.JavalinJackson;
